pyproject.toml
[tool.poetry]
name = "modul_11_homework"
version = "0.1.0"
description = "REST API to manage contacts using FastAPI and SQLAlchemy"
authors = ["RyszardKozik <149445001+RyszardKozik@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.85.0"
sqlalchemy = "^1.4.0"
psycopg2-binary = "^2.9.0"
pydantic = "^1.9.0"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
uvicorn = "^0.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
